Fixes a bug when the same machine is added two times to the server list (Fixes issue 419)
The problem was in gebr/src/gebr/server.c:232, with the 'i_name' variable
which was being NULL, crashing the `strcmp' function. We should investigate
why this value was NULL in the scenario, where the same machine was added
two times to the server list ('localhost' and 'maresias', for instance).