Help validation icon is updated when the help is saved (Fixes issue 423)
The problem was happening because the template help was being introduced in
the XML of the menu/program to be edited. This caused the program to think
the help was set, but if the user canceled the edition, the help was reverted
and the validation was not updated.

The function for editing help, now requires an extra parameter which is the
initial help string, avoiding the need to set the help in the XML structure.