Added a 'host' key to GeBRd's configuration file
Update issue 328

Added a 'host' key to GeBRd's configuration file. This is a comma separated
list of addresses that will be inserted in the MPI command line. For now,
there is only the OpenMPI implementation, which means the hosts are added with
the '--host' flag.
In case there is no host, nothing is added to the command line.