Changed parent type for HelpEditWindow from GtkWindow to GtkDialog
This change is necessary because GeBR uses a lot of modal dialogs. Since modal
dialogs prevent GtkWindow's from receiving events, we need to make this class
extend GtkDialog so it creates its own main loop, enabling it to receive
events. To call present this window, you should run it with gtk_dialog_run()
and then destroy with gtk_widget_destroy().