libgebr: validator: add gebr_validator_push/pop_document
Deprecates gebr_validator_set_document() in favor of the new methods
gebr_validator_push_document and gebr_validator_pop_document. This is
beneficial because often there is need for returning to the last state after a
gebr_validator_set_document(). Doing this way there is no need for knowing what
was the last document registered in the validator.

Example 1.

Code like this one
    gebr_validator_set_document(validator, &flow, GEBR_GEOXML_TYPE_FLOW, FALSE);
    // ... code
    gebr_validator_set_document(validator, &gebr.flow, GEBR_GEOXML_TYPE_FLOW, FALSE);

Should become like this
    gebr_validator_push_document(validator, &flow, GEBR_GEOXML_TYPE_FLOW);
    // ... code
    gebr_validator_pop_document(validator, GEBR_GEOXML_TYPE_FLOW);